+++ src/main/kotlin/icu/windea/pls/lang/injection/ParadoxScriptLanguageInjector.kt
@@ -8,11 +8,10 @@ import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiErrorElement
 import com.intellij.psi.PsiLanguageInjectionHost
-import com.intellij.psi.util.elementType
 import com.intellij.util.InjectionUtils
 import icu.windea.pls.core.castOrNull
 import icu.windea.pls.core.collections.orNull
-import icu.windea.pls.core.processChild
+import icu.windea.pls.core.orNull
 import icu.windea.pls.core.runCatchingCancelable
 import icu.windea.pls.core.unquote
 import icu.windea.pls.ep.configContext.ParameterValueCwtConfigContextProvider
@@ -20,15 +19,16 @@ import icu.windea.pls.ep.parameter.ParadoxParameterSupport
 import icu.windea.pls.lang.psi.mock.ParadoxParameterElement
 import icu.windea.pls.lang.util.ParadoxParameterManager
 import icu.windea.pls.model.ParadoxParameterContextReferenceInfo
+import icu.windea.pls.model.ParadoxSeparatorType
+import icu.windea.pls.model.constants.PlsPatternConstants
 import icu.windea.pls.model.injection.ParadoxParameterValueInjectionInfo
 import icu.windea.pls.script.ParadoxScriptLanguage
 import icu.windea.pls.script.psi.ParadoxParameter
-import icu.windea.pls.script.psi.ParadoxScriptElementTypes
 import icu.windea.pls.script.psi.ParadoxScriptString
-import icu.windea.pls.script.psi.ParadoxScriptStringExpressionElement
+import icu.windea.pls.script.psi.propertyKey
 
 /**
- * 脚本语言的语言注入器，用于提供以下功能：
+ * 脚本语言的语言注入器。
  *
  * * 对脚本参数的传入值进行语言注入（注入为脚本片段），以便推断对应的CWT规则上下文，从而提供高级语言功能。
  * * 对脚本参数的默认值进行语言注入（注入为脚本片段），以便推断对应的CWT规则上下文，从而提供高级语言功能。
@@ -37,12 +37,12 @@ import icu.windea.pls.script.psi.ParadoxScriptStringExpressionElement
  * @see ParameterValueCwtConfigContextProvider
  */
 class ParadoxScriptLanguageInjector : MultiHostInjector {
-    //see: com.intellij.util.InjectionUtils
-    //see: com.intellij.psi.impl.source.tree.injected.InjectedFileViewProvider
-    //see: org.intellij.plugins.intelliLang.inject.InjectorUtils
+    // see: com.intellij.util.InjectionUtils
+    // see: com.intellij.psi.impl.source.tree.injected.InjectedFileViewProvider
+    // see: org.intellij.plugins.intelliLang.inject.InjectorUtils
 
     private val toInject = listOf(
-        ParadoxScriptStringExpressionElement::class.java,
+        ParadoxScriptString::class.java,
         ParadoxParameter::class.java
     )
 
@@ -52,7 +52,7 @@ class ParadoxScriptLanguageInjector : MultiHostInjector {
 
     override fun getLanguagesToInject(registrar: MultiHostRegistrar, host: PsiElement) {
         if (host !is PsiLanguageInjectionHost) return
-        InjectionUtils.enableInjectLanguageAction(host, false) //disable inject language action
+        InjectionUtils.enableInjectLanguageAction(host, false) // disable inject language action
 
         runCatchingCancelable {
             doGetLanguageToInject(host, registrar)
@@ -60,7 +60,7 @@ class ParadoxScriptLanguageInjector : MultiHostInjector {
     }
 
     private fun doGetLanguageToInject(host: PsiLanguageInjectionHost, registrar: MultiHostRegistrar): Boolean {
-        if (host.lastChild is PsiErrorElement) return true //skip if host has syntax error
+        if (host.lastChild is PsiErrorElement) return true // skip if host has syntax error
 
         val allInjectionInfos = mutableListOf<ParadoxParameterValueInjectionInfo>()
 
@@ -69,6 +69,7 @@ class ParadoxScriptLanguageInjector : MultiHostInjector {
         ProgressManager.checkCanceled()
         applyInjectionForParameterDefaultValue(host, allInjectionInfos)
         host.putUserData(ParadoxParameterManager.Keys.parameterValueInjectionInfos, allInjectionInfos.orNull())
+
         if (allInjectionInfos.isEmpty()) return true
         allInjectionInfos.forEach f@{ injectionInfo ->
             registrar.startInjecting(ParadoxScriptLanguage)
@@ -80,26 +81,35 @@ class ParadoxScriptLanguageInjector : MultiHostInjector {
 
     private fun applyInjectionForArgumentValue(host: PsiElement, allInjectionInfos: MutableList<ParadoxParameterValueInjectionInfo>) {
         if (host !is ParadoxScriptString) return
-        val injectionInfos = getInjectionInfosForArgumentValue(host)
-        if (injectionInfos.isEmpty()) return
+        val argumentName = host.propertyKey?.name?.orNull() ?: return // 排除参数名不存在或为空的情况
+        val argumentValue = host.text.orNull() ?: return // 参数参数值为空的情况
+        if (!acceptInjectionForArgumentValue(host, argumentName, argumentValue)) return // 排除其他一些特殊情况
+        val injectionInfos = getInjectionInfosForArgumentValue(host, argumentValue)
         allInjectionInfos.addAll(injectionInfos)
     }
 
-    private fun getInjectionInfosForArgumentValue(host: ParadoxScriptString): List<ParadoxParameterValueInjectionInfo> {
-        //这里先向上得到contextReferenceInfo，接着获取传入值对应的textRange，然后选用在host的textRange之内的那些
+    @Suppress("unused")
+    private fun acceptInjectionForArgumentValue(host: ParadoxScriptString, argumentName: String, argumentValue: String): Boolean {
+        if (!PlsPatternConstants.parameterName.matches(argumentName)) return false // 参数名必须合法
+        val trimmed = argumentValue.unquote().trim()
+        if (ParadoxSeparatorType.entries.any { it.id == trimmed }) return false // 为一些狡猾人行方便
+        return true
+    }
+
+    private fun getInjectionInfosForArgumentValue(host: ParadoxScriptString, argumentValue: String): List<ParadoxParameterValueInjectionInfo> {
+        // 这里先向上得到contextReferenceInfo，接着获取传入值对应的textRange，然后选用在host的textRange之内的那些
         val from = ParadoxParameterContextReferenceInfo.From.InContextReference
         val contextReferenceInfo = ParadoxParameterSupport.getContextReferenceInfo(host, from = from) ?: return emptyList()
         if (contextReferenceInfo.arguments.isEmpty()) return emptyList()
         val hostRange = host.textRange
         return contextReferenceInfo.arguments.mapNotNull t1@{ referenceInfo ->
-            //这里需要特殊处理传入参数值被双引号括起的情况
+            // 这里需要特殊处理传入参数值被双引号括起的情况
             val rawRangeInsideHost = referenceInfo.argumentValueRange
                 ?.takeIf { it.startOffset >= hostRange.startOffset && it.endOffset <= hostRange.endOffset }
                 ?.shiftLeft(hostRange.startOffset)
                 ?: return@t1 null
-            val text = host.text
-            val rangeInsideHost = rawRangeInsideHost.unquote(rawRangeInsideHost.substring(text))
-            //这里要求参数值两边都有双引号
+            val rangeInsideHost = rawRangeInsideHost.unquote(rawRangeInsideHost.substring(argumentValue))
+            // 这里要求参数值两边都有双引号
             val parameterValueQuoted = rawRangeInsideHost.startOffset != rangeInsideHost.startOffset && rawRangeInsideHost.endOffset != rangeInsideHost.endOffset
             val parameterElementProvider = lazy {
                 val argumentNameElement = referenceInfo.argumentNameElement ?: return@lazy null
@@ -113,39 +123,32 @@ class ParadoxScriptLanguageInjector : MultiHostInjector {
                     reference.resolve()?.castOrNull<ParadoxParameterElement>()
                 }
             }
-            val textToInject = rangeInsideHost.substring(text)
+            val textToInject = rangeInsideHost.substring(argumentValue)
             ParadoxParameterValueInjectionInfo(textToInject, rangeInsideHost, parameterValueQuoted, parameterElementProvider)
         }
     }
 
     private fun applyInjectionForParameterDefaultValue(host: PsiElement, allInjectionInfos: MutableList<ParadoxParameterValueInjectionInfo>) {
         if (host !is ParadoxParameter) return
-        val injectionInfo = getInjectionInfoForParameterDefaultValue(host)
-        if (injectionInfo == null) return
+        val parameterName = host.name?.orNull() ?: return // 排除参数名不存在或为空的情况
+        val defaultValue = host.defaultValue?.orNull() ?: return // 排除默认值不存在或为空的情况
+        if (!acceptInjectionForParameterDefaultValue(host, parameterName, defaultValue)) return // 排除其他一些特殊情况
+        val injectionInfo = getInjectionInfoForParameterDefaultValue(host, defaultValue)
         allInjectionInfos.add(injectionInfo)
     }
 
-    private fun getInjectionInfoForParameterDefaultValue(host: ParadoxParameter): ParadoxParameterValueInjectionInfo? {
-        val parameterName = host.name
-        if (parameterName.isNullOrEmpty()) return null //skip if host is invalid
-        var start = -1
-        var end = -1
-        host.processChild { e ->
-            val elementType = e.elementType
-            if (elementType == ParadoxScriptElementTypes.PIPE) {
-                start = e.startOffsetInParent + 1
-            } else if (elementType == ParadoxScriptElementTypes.PARAMETER_END && start != -1) {
-                end = e.startOffsetInParent
-            }
-            true
-        }
-        if (start == -1 || end == -1) return null
-        val rangeInsideHost = TextRange.create(start, end)
-        val parameterValueQuoted = false
-        val parameterElementProvider = lazy {
-            ParadoxParameterSupport.resolveParameter(host)
-        }
+    @Suppress("unused")
+    private fun acceptInjectionForParameterDefaultValue(host: ParadoxParameter, parameterName: String, defaultValue: String): Boolean {
+        val trimmed = defaultValue.unquote().trim()
+        if (ParadoxSeparatorType.entries.any { it.id == trimmed }) return false // 为一些狡猾人行方便
+        return true
+    }
+
+    private fun getInjectionInfoForParameterDefaultValue(host: ParadoxParameter, defaultValue: String): ParadoxParameterValueInjectionInfo {
+        val defaultValueIndex = host.text.indexOf(defaultValue)
+        val rangeInsideHost = TextRange.from(defaultValueIndex, defaultValue.length)
+        val parameterElementProvider = lazy { ParadoxParameterSupport.resolveParameter(host) }
         val textToInject = rangeInsideHost.substring(host.text)
-        return ParadoxParameterValueInjectionInfo(textToInject, rangeInsideHost, parameterValueQuoted, parameterElementProvider)
+        return ParadoxParameterValueInjectionInfo(textToInject, rangeInsideHost, false, parameterElementProvider)
     }
 }